+++ custom_components/strato_dyndns/const.py
@@ -4,6 +4,7 @@ CONF_ACCOUNT_NAME = "account_name"
 CONF_DOMAINS = "domains"
 CONF_UPDATE_INTERVAL = "update_interval"
 CONF_IPV6_ENABLED = "ipv6_enabled"
+CONF_NOTIFICATIONS_ENABLED = "notifications_enabled"
 
 DEFAULT_UPDATE_INTERVAL = 30  # seconds